Added changes

+++ src/main/java/iot/presentation/controllers/SensorController.java
@@ -1,5 +1,6 @@
 package iot.presentation.controllers;
 
+import iot.utils.AuthUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -13,6 +14,9 @@ import iot.presentation.transport.SensorDTO;
 import iot.presentation.transport.SensorValueDTO;
 import net.minidev.json.JSONObject;
 
+import javax.servlet.http.HttpSession;
+import java.util.List;
+
 @Controller
 @RequestMapping("/sensor")
 public class SensorController {
@@ -22,7 +26,7 @@ public class SensorController {
 
 	@ResponseBody
 	@RequestMapping(value = "/{sensorId}", method = RequestMethod.GET)
-	public JSONObject getSensor(@RequestParam long sensorId) {
+	public JSONObject getSensor(@RequestParam long sensorId, HttpSession session) {
 		JSONObject object = new JSONObject();
 		SensorDTO sensor = sensorService.getSensorById(sensorId);
 
@@ -33,37 +37,61 @@ public class SensorController {
 
 	@ResponseBody
 	@RequestMapping(value = "/addSensor", method = RequestMethod.POST)
-	public Boolean addSensor(@RequestBody SensorDTO sensor) {
-		Boolean result = sensorService.addSensor(sensor);
+	public JSONObject addSensor(@RequestBody SensorDTO sensor, HttpSession session) {
+		JSONObject obj = new JSONObject();
 
-		return result;
+		if(AuthUtils.isUserAuthenticated(session)){
+			sensor.setId(null);
+			Boolean result = sensorService.addSensor(sensor);
+			obj.put("statusText", result?"success":"error");
+		}
+
+		return obj;
 	}
 
 	@ResponseBody
 	@RequestMapping(value = "/removeSensor", method = RequestMethod.POST)
-	public Boolean removeSensor(@RequestParam long sensorId) {
+	public JSONObject removeSensor(@RequestParam long sensorId, HttpSession session) {
+		JSONObject obj = new JSONObject();
+		if(AuthUtils.isUserAuthenticated(session)){
+			Boolean result = sensorService.removeSensor(sensorId);
 
-		Boolean result = sensorService.removeSensor(sensorId);
+			obj.put("statusText", result);
+		}
 
-		return result;
+
+		return obj;
 	}
 
 	@ResponseBody
 	@RequestMapping(value = "/editSensor", method = RequestMethod.POST)
-	public Boolean removeSensor(@RequestBody SensorDTO sensor) {
+	public Boolean removeSensor(@RequestBody SensorDTO sensor, HttpSession session) {
+		Boolean result = false;
 
-		Boolean result = sensorService.editSensor(sensor);
+		if(AuthUtils.isUserAuthenticated(session)){
+			 result = sensorService.editSensor(sensor);
+		}
 
 		return result;
 	}
 	
 	@ResponseBody
 	@RequestMapping(value = "/addSensorValue", method = RequestMethod.POST)
-	public Boolean addSensorValue(@RequestBody SensorValueDTO sensorValue) {
+	public Boolean addSensorValue(@RequestBody SensorValueDTO sensorValue, HttpSession session) {
 
 		Boolean result = sensorService.addSensorValue(sensorValue);
 
 		return result;
+	}
+
+	@ResponseBody
+	@RequestMapping(value = "/getSensorData", method = RequestMethod.POST)
+	public JSONObject addSensor(@RequestParam long sensorId, HttpSession session) {
+		JSONObject jsonObject = new JSONObject();
+		List<SensorValueDTO> sensorValues= sensorService.getSensorValues(sensorId);
+		jsonObject.put("values",sensorValues);
+
+		return jsonObject;
 	}
 
 }
